Revert "add: video 테스트 코드 작성"
This reverts commit 63d048f3397ec9a28045bda2d31c597eda5bd00d.

+++ src/video/video.service.ts
@@ -1,27 +1,20 @@
-import { Injectable, NotFoundException, UnauthorizedException } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { VideoEntity } from './entities/video.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { channel } from 'diagnostics_channel';
 import { VideoDto } from './dto/video.dto';
 import { Visibility } from './video.visibility.enum';
-import { UpdateVideoDto } from './dto/update.video.dto';
-import { UserEntity } from '../user/entity/user.entity';
-import { ChannelEntity } from '../channel/entities/channel.entity';
 
 @Injectable()
 export class VideoService {
   constructor(
     @InjectRepository(VideoEntity)
     private videoRepository: Repository<VideoEntity>,
-    @InjectRepository(ChannelEntity)
-    private channelRepository: Repository<ChannelEntity>,
   ) {}
 
-  async createVideo(user: UserEntity, videoDto: VideoDto): Promise<VideoEntity> {
+  async createVideo(videoDto: VideoDto): Promise<VideoEntity> {
     const { title, description, thumbnailURL, hashtags, duration, visibility } = videoDto;
-
-    const foundChannel = await this.findChannelByUserId(user.id);
 
     const video = this.videoRepository.create({
       title,
@@ -30,7 +23,6 @@ export class VideoService {
       hashtags,
       duration,
       visibility,
-      channel: foundChannel,
     });
 
     const saveVideo = this.videoRepository.save(video);
@@ -54,14 +46,13 @@ export class VideoService {
     return foundVideo;
   }
 
-  async updateVideo(
-    user: UserEntity,
-    videoId: number,
-    updateVideoDto: UpdateVideoDto,
-  ): Promise<VideoEntity> {
-    await this.findChannelByUserId(user.id);
-
-    await this.findVideoById(videoId);
+  async updateVideo(videoId: number, updateVideoDto): Promise<VideoEntity> {
+    const foundVideo = await this.videoRepository.findOne({
+      where: { id: videoId },
+    });
+    if (!foundVideo) {
+      throw new NotFoundException('존재 하지 않는 비디오 입니다.');
+    }
 
     const updateData = await this.updateDetails(updateVideoDto);
 
@@ -71,10 +62,13 @@ export class VideoService {
     return updatedVideo;
   }
 
-  async deleteVideo(user: UserEntity, videoId: number): Promise<object> {
-    await this.findChannelByUserId(user.id);
-
-    await this.findVideoById(videoId);
+  async deleteVideo(videoId: number): Promise<object> {
+    const foundVideo = await this.videoRepository.findOne({
+      where: { id: videoId },
+    });
+    if (!foundVideo) {
+      throw new NotFoundException('존재 하지 않는 비디오 입니다.');
+    }
 
     await this.videoRepository.delete({
       id: videoId,
@@ -83,25 +77,7 @@ export class VideoService {
     return { message: '동영상이 삭제되었습니다.' };
   }
 
-  private async findChannelByUserId(id) {
-    const foundChannel = await this.channelRepository.findOne({ where: { userId: id } });
-    if (!foundChannel) {
-      throw new UnauthorizedException('채널이 존재하지 않습니다.');
-    }
-    return foundChannel;
-  }
-
-  private async findVideoById(id) {
-    const foundVideo = await this.videoRepository.findOne({
-      where: { id: id },
-    });
-    if (!foundVideo) {
-      throw new NotFoundException('존재 하지 않는 비디오 입니다.');
-    }
-    return foundVideo;
-  }
-
-  private async updateDetails(updateVideoDto: UpdateVideoDto) {
+  private async updateDetails(updateVideoDto) {
     const updateData: Partial<VideoEntity> = {};
 
     if (updateVideoDto.title) {
@@ -112,12 +88,8 @@ export class VideoService {
       updateData.description = updateVideoDto.description;
     }
 
-    if (updateVideoDto.thumbnailURL) {
-      updateData.thumbnailURL = updateVideoDto.thumbnailURL;
-    }
-
-    if (updateVideoDto.hashtags) {
-      updateData.hashtags = updateVideoDto.hashtags;
+    if (updateVideoDto.thumbnail_url) {
+      updateData.description = updateVideoDto.thumbnail_url;
     }
 
     if (updateVideoDto.visibility) {
